# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: VeriHire — verified workforce management API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /workers:
    get:
      operationId: listWorkers
      tags: [workers]
      summary: List workers (filterable)
      parameters:
        - in: query
          name: category
          schema: { type: string }
        - in: query
          name: city
          schema: { type: string }
        - in: query
          name: verificationStatus
          schema:
            type: string
            enum: [pending, approved, rejected]
        - in: query
          name: minExperience
          schema: { type: integer }
        - in: query
          name: maxMonthlySalary
          schema: { type: integer }
        - in: query
          name: maxHourlyRate
          schema: { type: integer }
        - in: query
          name: search
          schema: { type: string }
      responses:
        "200":
          description: List of workers
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Worker" }
    post:
      operationId: createWorker
      tags: [workers]
      summary: Create a worker profile
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/WorkerInput" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Worker" }
...
  /workers/{id}:
    get:
      operationId: getWorker
      tags: [workers]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Worker
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Worker" }
        "404":
          description: Not found
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
    patch:
      operationId: updateWorker
      tags: [workers]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/WorkerUpdate" }
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Worker" }
  /workers/{id}/verify:
    post:
      operationId: verifyWorker
      tags: [workers, admin]
      summary: Approve or reject a worker (admin)
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/VerifyWorkerInput" }
      responses:
        "200":
          description: Updated worker
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Worker" }
  /workers/{id}/reviews:
    get:
      operationId: listWorkerReviews
      tags: [reviews]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Reviews for worker
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Review" }

  # ---------- Customers ----------
...
  /tasks:
    get:
      operationId: listTasks
      tags: [tasks]
      parameters:
        - in: query
          name: status
          schema:
            type: string
            enum: [open, accepted, in_progress, completed, cancelled]
        - in: query
          name: category
          schema: { type: string }
        - in: query
          name: city
          schema: { type: string }
        - in: query
          name: customerId
          schema: { type: integer }
        - in: query
          name: workerId
          schema: { type: integer }
      responses:
        "200":
          description: Tasks
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Task" }
    post:
      operationId: createTask
      tags: [tasks]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/TaskInput" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Task" }
  /tasks/{id}:
    get:
      operationId: getTask
      tags: [tasks]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Task
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Task" }
        "404":
          description: Not found
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
  /tasks/{id}/accept:
    post:
      operationId: acceptTask
      tags: [tasks]
      summary: Worker accepts an open task
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/AcceptTaskInput" }
      responses:
        "200":
          description: Task accepted
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Task" }
  /tasks/{id}/status:
    patch:
      operationId: updateTaskStatus
      tags: [tasks]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/TaskStatusInput" }
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Task" }

  # ---------- Reviews ----------
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error: { type: string }
      required: [error]

    # ---------- Worker ----------
    Worker:
      type: object
      properties:
        id: { type: integer }
        fullName: { type: string }
        phone: { type: string }
        photoUrl:
          type: ["string", "null"]
        category: { type: string, description: "maid | driver | guard | cook | nanny | gardener | other" }
        skills:
          type: array
          items: { type: string }
        bio:
          type: ["string", "null"]
        experienceYears: { type: integer }
        expectedMonthlySalary:
          type: ["integer", "null"]
        hourlyRate:
          type: ["integer", "null"]
        city: { type: string }
        serviceArea:
          type: ["string", "null"]
        cnicNumber: { type: string }
        cnicImageUrl:
          type: ["string", "null"]
        verificationStatus:
          type: string
          enum: [pending, approved, rejected]
        rejectionReason:
          type: ["string", "null"]
        averageRating: { type: number }
        reviewCount: { type: integer }
        completedJobs: { type: integer }
        availability:
          type: string
          enum: [available, busy, offline]
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
      required:
        - id
        - fullName
        - phone
        - category
        - skills
        - experienceYears
        - city
        - cnicNumber
        - verificationStatus
        - averageRating
        - reviewCount
        - completedJobs
        - availability
        - createdAt
        - updatedAt
    WorkerInput:
      type: object
      properties:
        fullName: { type: string }
        phone: { type: string }
        photoUrl: { type: ["string", "null"] }
        category: { type: string }
        skills:
          type: array
          items: { type: string }
        bio: { type: ["string", "null"] }
        experienceYears: { type: integer }
        expectedMonthlySalary: { type: ["integer", "null"] }
        hourlyRate: { type: ["integer", "null"] }
        city: { type: string }
        serviceArea: { type: ["string", "null"] }
        cnicNumber: { type: string }
        cnicImageUrl: { type: ["string", "null"] }
      required:
        - fullName
        - phone
        - category
        - skills
        - experienceYears
        - city
        - cnicNumber
    WorkerUpdate:
      type: object
      properties:
        fullName: { type: string }
        phone: { type: string }
        photoUrl: { type: ["string", "null"] }
        category: { type: string }
        skills:
          type: array
          items: { type: string }
        bio: { type: ["string", "null"] }
        experienceYears: { type: integer }
        expectedMonthlySalary: { type: ["integer", "null"] }
        hourlyRate: { type: ["integer", "null"] }
        city: { type: string }
        serviceArea: { type: ["string", "null"] }
        availability:
          type: string
          enum: [available, busy, offline]
    VerifyWorkerInput:
      type: object
      properties:
        decision:
          type: string
          enum: [approved, rejected]
        rejectionReason: { type: ["string", "null"] }
      required: [decision]

    # ---------- Customer ----------
    Customer:
      type: object
      properties:
        id: { type: integer }
        fullName: { type: string }
        phone: { type: string }
        city: { type: string }
        photoUrl: { type: ["string", "null"] }
        createdAt: { type: string, format: date-time }
      required: [id, fullName, phone, city, createdAt]
...
    Task:
      type: object
      properties:
        id: { type: integer }
        customerId: { type: integer }
        title: { type: string }
        description: { type: string }
        category: { type: string }
        durationHours: { type: integer }
        location: { type: string }
        city: { type: string }
        budget: { type: integer }
        status:
          type: string
          enum: [open, accepted, in_progress, completed, cancelled]
        scheduledFor: { type: ["string", "null"], format: date-time }
        acceptedByWorkerId: { type: ["integer", "null"] }
        acceptedAt: { type: ["string", "null"], format: date-time }
        completedAt: { type: ["string", "null"], format: date-time }
        customer: { $ref: "#/components/schemas/Customer" }
        worker: { $ref: "#/components/schemas/Worker" }
        createdAt: { type: string, format: date-time }
      required:
        - id
        - customerId
        - title
        - description
        - category
        - durationHours
        - location
        - city
        - budget
        - status
        - createdAt
    TaskInput:
      type: object
      properties:
        customerId: { type: integer }
        title: { type: string }
        description: { type: string }
        category: { type: string }
        durationHours: { type: integer }
        location: { type: string }
        city: { type: string }
        budget: { type: integer }
        scheduledFor: { type: ["string", "null"], format: date-time }
      required:
        - customerId
        - title
        - description
        - category
        - durationHours
        - location
        - city
        - budget
    AcceptTaskInput:
      type: object
      properties:
        workerId: { type: integer }
      required: [workerId]
    TaskStatusInput:
      type: object
      properties:
        status:
          type: string
          enum: [open, accepted, in_progress, completed, cancelled]
      required: [status]

    # ---------- Review ----------
    Review:
      type: object
      properties:
        id: { type: integer }
        workerId: { type: integer }
        customerId: { type: integer }
        bookingId: { type: ["integer", "null"] }
        taskId: { type: ["integer", "null"] }
        rating: { type: integer }
        comment: { type: ["string", "null"] }
        customer: { $ref: "#/components/schemas/Customer" }
        createdAt: { type: string, format: date-time }
      required: [id, workerId, customerId, rating, createdAt]
...