# lib/api-zod/src/generated/types/workerInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * VeriHire — verified workforce management API
 * OpenAPI spec version: 0.1.0
 */

export interface WorkerInput {
  fullName: string;
  phone: string;
  /** @nullable */
  photoUrl?: string | null;
  category: string;
  skills: string[];
  /** @nullable */
  bio?: string | null;
  experienceYears: number;
  /** @nullable */
  expectedMonthlySalary?: number | null;
  /** @nullable */
  hourlyRate?: number | null;
  city: string;
  /** @nullable */
  serviceArea?: string | null;
  cnicNumber: string;
  /** @nullable */
  cnicImageUrl?: string | null;
}